+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add fileLoader option to allow specify file extension explicitly.
 
 ## [1.2.0] - 2017-10-17
 ### Added